Swagger 3 [bugs fixed] + password encoding

+++ web/src/main/java/com.luchkovskiy/controllers/requests/update/AccidentUpdateRequest.java
@@ -21,7 +21,7 @@ public class AccidentUpdateRequest extends AccidentCreateRequest {
 
     @NotNull
     @Min(1)
-    @Schema(requiredMode = Schema.RequiredMode.REQUIRED, example = "1", type = "Long", description = "Id of the accident")
+    @Schema(requiredMode = Schema.RequiredMode.REQUIRED, example = "1", type = "integer", description = "Id of the accident")
     private Long id;
 
 }
