+++ web/src/main/java/com.luchkovskiy/controllers/requests/update/CarUpdateRequest.java
@@ -21,7 +21,7 @@ public class CarUpdateRequest extends CarCreateRequest {
 
     @NotNull
     @Min(1)
-    @Schema(requiredMode = Schema.RequiredMode.REQUIRED, example = "1", type = "Long", description = "Id of the car")
+    @Schema(requiredMode = Schema.RequiredMode.REQUIRED, example = "1", type = "integer", description = "Id of the car")
     private Long id;
 
 }
