+++ web/src/main/java/com.luchkovskiy/controllers/requests/update/SessionUpdateRequest.java
@@ -21,7 +21,7 @@ public class SessionUpdateRequest extends SessionCreateRequest {
 
     @NotNull
     @Min(1)
-    @Schema(requiredMode = Schema.RequiredMode.REQUIRED, example = "1", type = "Long", description = "Id of the session")
+    @Schema(requiredMode = Schema.RequiredMode.REQUIRED, example = "1", type = "integer", description = "Id of the session")
     private Long id;
 
 }
