+++ web/src/main/java/com.luchkovskiy/controllers/requests/update/SubscriptionLevelUpdateRequest.java
@@ -21,7 +21,7 @@ public class SubscriptionLevelUpdateRequest extends SubscriptionLevelCreateReque
 
     @NotNull
     @Min(1)
-    @Schema(requiredMode = Schema.RequiredMode.REQUIRED, example = "1", type = "Long", description = "Id of the subscription level")
+    @Schema(requiredMode = Schema.RequiredMode.REQUIRED, example = "1", type = "integer", description = "Id of the subscription level")
     private Long id;
 
 }
