+++ web/src/main/java/com.luchkovskiy/controllers/requests/update/SubscriptionUpdateRequest.java
@@ -21,7 +21,7 @@ public class SubscriptionUpdateRequest extends SubscriptionCreateRequest {
 
     @NotNull
     @Min(1)
-    @Schema(requiredMode = Schema.RequiredMode.REQUIRED, example = "1", type = "Long", description = "Id of the subscription")
+    @Schema(requiredMode = Schema.RequiredMode.REQUIRED, example = "1", type = "integer", description = "Id of the subscription")
     private Long id;
 
 }
